+++ StarlingRoundUpChallenge/Models/StarlingApi/CurrencyAndAmount.cs
@@ -3,5 +3,5 @@ namespace StarlingRoundUpChallenge.Models.StarlingApi;
 public class CurrencyAndAmount
 {
     public string Currency { get; set; }
-    public int MinorUnits { get; set; }
+    public long MinorUnits { get; set; }
 }
